refactor: embed vexflow-repl directly in project
Up until this point the library which I created for rendering was
separate from this project since I thought it shouldn't be limited to
the scope of this project.

Because I didn't use submodules and worked on both repos things got out
of hand so I reverted a bit and brought the code from the lib here.

In the future I'll have to "unembed" this lib and use submodules if its
functionality is needed somewhere else.